Separate non-Windows platform detection to its own file to fix ilc errors
