+++ src/System.Globalization/tests/Configurations.props
@@ -3,6 +3,7 @@
   <PropertyGroup>
     <BuildConfigurations>
       netstandard;
+      uap-Windows_NT;
     </BuildConfigurations>
   </PropertyGroup>
 </Project>
