+++ src/System.Globalization/tests/DateTimeFormatInfo/DateTimeFormatInfoData.cs
@@ -29,19 +29,23 @@ namespace System.Globalization.Tests
 
         public static string[] FrFRDayNames()
         {
+#if !uap
             if (PlatformDetection.IsOSX && PlatformDetection.OSXKernelVersion < new Version(16, 0))
             {
                 return new string[] { "Dimanche", "Lundi", "Mardi", "Mercredi", "Jeudi", "Vendredi", "Samedi" };
             }
+#endif
             return new string[] { "dimanche", "lundi", "mardi", "mercredi", "jeudi", "vendredi", "samedi" };
         }
 
         public static string[] FrFRAbbreviatedDayNames()
         {
+#if !uap
             if (PlatformDetection.IsOSX  && PlatformDetection.OSXKernelVersion < new Version(16, 0))
             {
                 return new string[] { "Dim.", "Lun.", "Mar.", "Mer.", "Jeu.", "Ven.", "Sam." };
             }
+#endif
             return new string[] { "dim.", "lun.", "mar.", "mer.", "jeu.", "ven.", "sam." };
         }
